Создано приложение Блог, весь функционал работы с моделью Блог и сама модель перенесены в новое приложение
