+++ blog/templates/blog/blog_confirm_delete.html
@@ -19,7 +19,7 @@
             {% csrf_token %}
             <div class="row">
               <button class="btn btn-danger col-4 m-1" type="submit">Удалить</button>
-              <a class="btn btn-secondary col-4 m-1" href="{% url 'catalog:blogs' %}">Назад</a>
+              <a class="btn btn-secondary col-4 m-1" href="{% url 'blog:blogs' %}">Назад</a>
             </div>
           </form>
         </div>
