+++ blog/templates/blog/blog_detail.html
@@ -17,7 +17,7 @@
           <p class="card-text">{{ object.body }}</p>
           <p>Просмотров: {{ object.views_count }}</p>
           <div class="row">
-            <a class="btn btn-secondary col-4 m-1" href="{% url 'catalog:blogs' %}">Назад</a>
+            <a class="btn btn-secondary col-4 m-1" href="{% url 'blog:blogs' %}">Назад</a>
           </div>
         </div>
       </div>
