+++ blog/urls.py
@@ -1,0 +1,14 @@
+from django.urls import path
+
+from blog.apps import BlogConfig
+from blog.views import BlogListView, BlogCreateView, BlogDetailView, BlogUpdateView, BlogDeleteView
+
+app_name = BlogConfig.name
+
+urlpatterns = [
+    path('blog/create', BlogCreateView.as_view(), name='blog_create'),
+    path('blogs', BlogListView.as_view(), name='blogs'),
+    path('blog/detail/<int:pk>', BlogDetailView.as_view(), name='blog_detail'),
+    path('blog/update/<int:pk>', BlogUpdateView.as_view(), name='blog_update'),
+    path('blog/delete/<int:pk>', BlogDeleteView.as_view(), name='blog_delete'),
+]
