+++ catalog/urls.py
@@ -1,8 +1,8 @@
 from django.urls import path
 from catalog.apps import CatalogConfig
 from catalog.views import HomeTemplateView, ConsumableDetailView, ConsumablesListView, ContactsTemplateView, \
-    EquipmentListView, ConsumableCreateView, ConsumableUpdateView, BlogListView, BlogDetailView, BlogCreateView, \
-    BlogUpdateView, ConsumableDeleteView, EquipmentDetailView, BlogDeleteView, consumable_purchase_count
+    EquipmentListView, ConsumableCreateView, ConsumableUpdateView, ConsumableDeleteView, EquipmentDetailView, \
+    consumable_purchase_count
 
 app_name = CatalogConfig.name
 
@@ -16,10 +16,5 @@ urlpatterns = [
     path('consumable/create', ConsumableCreateView.as_view(), name='consumable_create'),
     path('consumable/<int:pk>/update', ConsumableUpdateView.as_view(), name='consumable_update'),
     path('consumable/<int:pk>/delete', ConsumableDeleteView.as_view(), name='consumable_delete'),
-    path('blogs', BlogListView.as_view(), name='blogs'),
-    path('blog/<int:pk>/detail', BlogDetailView.as_view(), name='blog_detail'),
-    path('blog/create', BlogCreateView.as_view(), name='blog_create'),
-    path('blog/<int:pk>/update', BlogUpdateView.as_view(), name='blog_update'),
-    path('blog/<int:pk>/delete', BlogDeleteView.as_view(), name='blog_delete'),
     path('consumable/<int:pk>/purchase', consumable_purchase_count, name='consumable_purchase'),
 ]
